pyproject.toml
[tool.poetry]
name = "project_chgk"
version = "0.1.0"
description = "Pet-project"
authors = ["Alexion <alex24bryant@mail.ru>"]
packages = [
    {include = 'chgk'}
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
